+++ bytestring/CHANGES.md
@@ -3,6 +3,7 @@
 ## Unreleased
 
 - Minimum supported Rust version (MSRV) is now 1.88.
+- Improve `From<ByteString> for String` performance.
 
 ## 1.5.0